Notes removed
Note that some of the translations aren't there and others also include the Saheeh International translation. You can find the other translations easily on Quran.com if you want to edit it but I've left them like this here. The tagging and formatting is still consistent with the other surahs.